Moved main parser_path loop to new TestCaseGenerator class.
The new class is going to do most of the top level heavy lifting for
looping things and calling the various components.  The aim is to have
better control of the order in which the things are called so that we
can do round-robins, etc.
